Adding two-way comparison of attributes, to allow for attributes that are not declared at all (vs having an undefined value) to be removed.
Capturing the existing attributes prior to update to find those that were removed by the update.
Adding virtual attributes declaration to make integration with conditional expressions in soy attribute declarations easier. The performance is not great, so it would be better to avoid it if possible.
Re-wording some test descriptions.