Changing alignment to temporarliy remove misplaced key items.
When the current node is a keyed item, remove it from the DOM instead of just moving the matching node before it. The node will be added back using it's entry in the key mapping if necessary. This reduces the number of move operations needed from O(n) to O(number of moves) when items are keyed. As a result, performance when keyed nodes are moved/removed from the start of a parent is greatly improved.

- Changed from completely removing the key map to simply removing the unused keys