Optimize createKeyMap for created elements
When it existed in the DOM beforehand, we'll need to check for any keyed
children.

But when we create an Element, there's no need to check it's children for
keyed elements, because it's brand new and we'll register any new
children as we encounter them. We can optimize for this case.