Optimize attributes diff in elementOpen
This optimizes the attributes diffing to only use `j` redundant access.
If a change is detected in the beginning (or near), then `j` is small
(possibly 0) and we get the largest gains. If the change is detected at
the end, then `j` is equal to the number of attribute name-values (so we
end up doing a second full iteration).

That's still better than the code before, which always guaranteed 2 full
iterations. Now we just do 1 full iteration, plus `j` (which may be
another full).

Time for bed. :sleeping::zzz::zzz: