Add optional typeId parameter.
The typeId parameter allows signaling to Incremental DOM that two statements are not equivalent. This allows a way of avoiding a transformation of a subtree into a completely different subtree. This can be helpful when a conditional subtree shares the same tag name as an adjacent call. For example:

```
if (condition) {
  open('div');
    text('Hello');
  close('div');
}
open('div');
  // Some large subtree
close('div');
```

Without a typeId, when the condition goes from `false` to `true`, the large subtree gets transformed into the text "Hello", removing all the DOM nodes. Afterwards, the large subtree is re-created. While keys can help accomplish the same goal, they may not suitable for code generation as an `open` call may exist in a function that is called multiple times within the same parent. While going from `true` to `false` currently recreates the large subtree, Incremental DOM may support handling a single deletion.

- Change core's `elementOpen` and `elementClose` to use `open` and `close`
- Expose `open` and `close`
- Add typeId to `open`