Change array truncation to use `pop` instead of setting length.
This change was originally proposed in #151 for the attributes array trunctation. While that is still generally a rare case, the `argsBuilder` (for `elementOpenStart`) is used quite often in practice.  The difference when changing the perf tests to use `elementOpenStart` is quite large with this change. This is also needed for future changes.