Do not track focus state on NodeData.
Remove the import of NodeData at the start of the patch for the focus
path, making inferred keys work correctly. The cost of checking the
focus path array is very small compared to either creating or moving a
Node.

Two side benefits are that we use less memory per Node and that we save
a few bytes on the gzipped payload.