When importing DOM, apply statics with different values. (#374)
When importing server rendered DOM, statics may not match when Elements
are reused. Ideally, a no-op patch should be applied (which would avoid
this problem), but it is not always feasible. Ideally, key should also
be used when statics are used, but some server-side rendered DOM might
omit keys (e.g. to reduce the payload transmitted).