Adding Towser, a framework written in Haxe
Hello! I would like to add my framework Towser to this doc. Towser is a framework that is heavily inspired by the Elm Architecture written in a strongly typed language called Haxe. The reason why I chose the incremental-dom for Towser is because I found (for my use-case) virtual-dom libraries to be too harsh on the GC. Towser is being developed to handle all UI in my game engine (which has many many updates) but also to be generic enough for any type of interactive website. 

I'm happy to provide more info. Thanks for the project, the api is a dream.