pin type of patchInner/patchOuter
Though createPatchInner etc are defined in terms of generics,
the resulting values saved here (patchInner/patchOuter) have
inexpressible types.  You'd want something like:
  const patchInner: <T>PatchFunction<T, Node>;
but that is not legal TS syntax.

What happens instead is in the d.ts, TS just chooses an
arbitrary T and all callers get that instead.  In TS3.4
the T is {} and in 3.5 it's unknown and that makes the
callers unhappy.  So this change pins this code to the
TS3.4 behavior.