Pierce ShadowRoot boundary when getting ancestry (#486)
* Pierce ShadowRoot boundary when getting ancestry

With on-demand ShadyDOM, the DOM is rendered without shadow nodes, but the API acts as if these shadow nodes exist. This can cause issues with getFocusedPath calls where the activeElement is inside of a "shadow node", thus getting the ancestry of the activeElement to the root results in an assertion error because `n.parentNode` returns undefined. We can get around this by piercing through the shadow node boundary.

This should only affect on-demand ShadyDOM since for regular ShadowDOM, the `!node.contains(activeElement)` check would be true causing getFocusedPath to short circuit.

* Remove unecessary optional chaining